refactor(tool): finish framework split
- move schema handling into core.tool.contracts
- move execution error reporting and source adaptation into core.tool.execution
- remove superseded framework modules and migrate all consumers
