+++ core/agent_harness/tools/action_tools.py
@@ -7,7 +7,7 @@ from typing import Any, Protocol
 
 from core.domain.types.tools import ToolSurface
 from core.tool.contracts import RegisteredTool
-from core.tool_framework.utils.integration_sources import availability_view
+from core.tool.execution import availability_view
 from platform.harness_ports import get_surface_tool_map, get_surface_tools
 from platform.observability.trace.redaction import redact_sensitive
