+++ integrations/azure/tools/azure_monitor_logs_tool/__init__.py
@@ -12,7 +12,7 @@ from config.constants.azure import (
     AZURE_MAX_RESULTS_HARD_LIMIT,
 )
 from core.domain.types.tools import ToolSurface
-from core.tool_framework.telemetry import report_run_error
+from core.tool.execution import report_run_error
 from core.tool_framework.tool_decorator import tool
 from core.tool_framework.utils.tool_availability import tool_unavailable
