+++ integrations/google_docs/tools/__init__.py
@@ -6,7 +6,7 @@ from __future__ import annotations
 
 from typing import Any
 
-from core.tool_framework.telemetry import report_run_error
+from core.tool.execution import report_run_error
 from core.tool_framework.tool_decorator import tool
 from integrations.config_models import GoogleDocsIntegrationConfig
 from integrations.google_docs.client import GoogleDocsClient
