+++ integrations/openclaw/tools/openclaw_mcp_tool/__init__.py
@@ -15,7 +15,7 @@ Package layout (separation of concerns):
 from __future__ import annotations
 
 from core.domain.types.tools import ToolSurface
-from core.tool_framework.telemetry import report_run_error
+from core.tool.execution import report_run_error
 from core.tool_framework.tool_decorator import tool
 from core.tool_framework.utils.mcp_tool_listing import build_mcp_tool_listing
 from integrations.openclaw import (
