+++ integrations/openobserve/tools/openobserve_logs_tool/__init__.py
@@ -9,7 +9,7 @@ from typing import Any
 import httpx
 
 from core.domain.types.tools import ToolSurface
-from core.tool_framework.telemetry import report_run_error
+from core.tool.execution import report_run_error
 from core.tool_framework.tool_decorator import tool
 from core.tool_framework.utils.tool_availability import tool_unavailable
