+++ integrations/slack/tools/slack_read_list_tool/tool.py
@@ -7,8 +7,7 @@ import re
 from typing import Any
 
 from core.domain.types.tools import ToolSurface
-from core.tool.contracts import BaseTool
-from core.tool.contracts import SideEffectLevel
+from core.tool.contracts import BaseTool, SideEffectLevel
 from core.tool_framework.tool_decorator import tool
 from core.tool_framework.utils.tool_availability import tool_unavailable
 from integrations.slack.tools.slack_read_list_tool.constants import (
