+++ integrations/slack/tools/slack_reply_message_tool/tool.py
@@ -5,8 +5,7 @@ from __future__ import annotations
 from typing import Any
 
 from core.domain.types.tools import ToolSurface
-from core.tool.contracts import BaseTool
-from core.tool.contracts import SideEffectLevel
+from core.tool.contracts import BaseTool, SideEffectLevel
 from core.tool_framework.tool_decorator import tool
 from integrations.slack.tools.slack_read_messages_tool.constants import SOURCE
 from integrations.slack.tools.slack_read_messages_tool.validation import validate_channel_id
