+++ tests/core/tool_framework/test_decorator.py
@@ -6,8 +6,7 @@ from typing import Any
 
 import pytest
 
-from core.tool.contracts import BaseTool
-from core.tool.contracts import REGISTERED_TOOL_ATTR, RegisteredTool
+from core.tool.contracts import REGISTERED_TOOL_ATTR, BaseTool, RegisteredTool
 from core.tool_framework.tool_decorator import tool
 
 # ---------------------------------------------------------------------------
