+++ tools/investigation/stages/gather_evidence/agent.py
@@ -28,8 +28,7 @@ from core.messages import MessageMapper, RuntimeMessage, ToolResultRuntimeMessag
 from core.provider import ProviderHooks, ProviderRequest
 from core.state import InvestigationState
 from core.state.evidence import EvidenceEntry
-from core.tool.contracts import RegisteredTool
-from core.tool.contracts import RuntimeTool
+from core.tool.contracts import RegisteredTool, RuntimeTool
 from platform.observability import debug_print
 from platform.observability import get_progress_tracker as get_tracker
 from platform.observability.trace.redaction import (
